backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1

# ML / XAI
scikit-learn==1.5.0
shap==0.45.1
umap-learn==0.5.6
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
networkx==3.3

# GCP / Google AI
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0
google-cloud-aiplatform==1.57.0
vertexai==1.57.0
langchain-google-vertexai==1.0.6

# LangGraph / LangChain
langgraph==0.1.14
langchain-core==0.2.10

# Clustering
hdbscan==0.8.38

# Testing
pytest==8.2.2
httpx==0.27.0
pytest-asyncio==0.23.7

# Utils
python-multipart==0.0.9
aiofiles==23.2.1
